removed the cyclic dependency
